pyproject.toml
[project]
name = "hexagonal-architecture-fastapi-back"
version = "0.1.0"
description = ""
authors = [{ name = "Santiago", email = "sasanchezramirez@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.123.0,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "dependency-injector (>=4.48.2,<5.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==3.2.2)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "email-validator (>=2.3.0,<3.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
